// MpulunguDay-main/types.ts
export interface Teacher {
  id: string;
  name: string;
  subject?: string; // Primary subject
  email: string;
  contact: string;
  status: 'Active' | 'On Leave';
  username?: string;
  password?: string;
}